added unit test for LocationData class
