+++ app/blueprints/assessments/templates/macros/application_overviews_table_cof.html
@@ -9,7 +9,9 @@
 {% endfor %}
 {%- endmacro %}
 
-{% macro render(application_overviews, round_details, query_params, asset_types, assessment_statuses,countries, regions, local_authorities, show_clear_filters, sort_column, sort_order, tag_option_groups, tags, tagging_purpose_config) -%}
+{% macro render(application_overviews, round_details, query_params, asset_types,
+   assessment_statuses,countries, regions, local_authorities, show_clear_filters,
+    sort_column, sort_order, tag_option_groups, tags, tagging_purpose_config) -%}
 <nav class="search-bar-flex-container">
   <form method="get" class="govuk-!-width-full">
     {{ filter_options(
